warning removal

+++ devl/nll/registration-ct-ct-planar.h
@@ -113,10 +113,10 @@ namespace algorithm
          {
             // computes the registration in YZ plane. We must multiply the bounding box by the spacing as the planes are formed like this...
             Registration2D registrationx;
-            Matrix tfmx = registrationx.compute( pxs, pxt, core::vector2i( minBoundingBoxSource[ 1 ] * source.getSpacing()[ 1 ], minBoundingBoxSource[ 2 ] * source.getSpacing()[ 2 ] ),
-                                                           core::vector2i( maxBoundingBoxSource[ 1 ] * source.getSpacing()[ 1 ], maxBoundingBoxSource[ 2 ] * source.getSpacing()[ 2 ] ),
-                                                           core::vector2i( minBoundingBoxTarget[ 1 ] * target.getSpacing()[ 1 ], minBoundingBoxTarget[ 2 ] * target.getSpacing()[ 2 ] ),
-                                                           core::vector2i( maxBoundingBoxTarget[ 1 ] * target.getSpacing()[ 1 ], maxBoundingBoxTarget[ 2 ] * target.getSpacing()[ 2 ] ) );
+            Matrix tfmx = registrationx.compute( pxs, pxt, core::vector2i( (int)( minBoundingBoxSource[ 1 ] * source.getSpacing()[ 1 ] ), (int)( minBoundingBoxSource[ 2 ] * source.getSpacing()[ 2 ] ) ),
+                                                           core::vector2i( (int)( maxBoundingBoxSource[ 1 ] * source.getSpacing()[ 1 ] ), (int)( maxBoundingBoxSource[ 2 ] * source.getSpacing()[ 2 ] ) ),
+                                                           core::vector2i( (int)( minBoundingBoxTarget[ 1 ] * target.getSpacing()[ 1 ] ), (int)( minBoundingBoxTarget[ 2 ] * target.getSpacing()[ 2 ] ) ),
+                                                           core::vector2i( (int)( maxBoundingBoxTarget[ 1 ] * target.getSpacing()[ 1 ] ), (int)( maxBoundingBoxTarget[ 2 ] * target.getSpacing()[ 2 ] ) ) );
 
             Matrix rotx( 4, 4 );
             rotx( 0, 0 ) = 1;
@@ -148,10 +148,10 @@ namespace algorithm
 
             // // computes the registration in XZ plane
             Registration2D registrationy;
-            Matrix tfmy = registrationy.compute( pys, pyt2, core::vector2i( minBoundingBoxSource[ 0 ] * source.getSpacing()[ 0 ], minBoundingBoxSource[ 2 ] * source.getSpacing()[ 2 ] ),
-                                                            core::vector2i( maxBoundingBoxSource[ 0 ] * source.getSpacing()[ 0 ], maxBoundingBoxSource[ 2 ] * source.getSpacing()[ 2 ] ),
-                                                            core::vector2i( minBoundingBoxTarget[ 0 ] * target.getSpacing()[ 0 ], minBoundingBoxTarget[ 2 ] * target.getSpacing()[ 2 ] ),
-                                                            core::vector2i( maxBoundingBoxTarget[ 0 ] * target.getSpacing()[ 0 ], maxBoundingBoxTarget[ 2 ] * target.getSpacing()[ 2 ] ) );
+            Matrix tfmy = registrationy.compute( pys, pyt2, core::vector2i( (int)( minBoundingBoxSource[ 0 ] * source.getSpacing()[ 0 ] ), (int)( minBoundingBoxSource[ 2 ] * source.getSpacing()[ 2 ] ) ),
+                                                            core::vector2i( (int)( maxBoundingBoxSource[ 0 ] * source.getSpacing()[ 0 ] ), (int)( maxBoundingBoxSource[ 2 ] * source.getSpacing()[ 2 ] ) ),
+                                                            core::vector2i( (int)( minBoundingBoxTarget[ 0 ] * target.getSpacing()[ 0 ] ), (int)( minBoundingBoxTarget[ 2 ] * target.getSpacing()[ 2 ] ) ),
+                                                            core::vector2i( (int)( maxBoundingBoxTarget[ 0 ] * target.getSpacing()[ 0 ] ), (int)( maxBoundingBoxTarget[ 2 ] * target.getSpacing()[ 2 ] ) ) );
 
             // do some postprocessing...
             Matrix roty( 4, 4 );
@@ -261,7 +261,7 @@ namespace algorithm
          if ( doTableRemoval )
             ymax = findTableY( pz );
          else
-            ymax = v.size()[ 1 ] * v.getSpacing()[ 1 ] - 1;
+            ymax = (int)( v.size()[ 1 ] * v.getSpacing()[ 1 ] ) - 1;
          if ( ymax > 0 )
          {
             #ifndef NLL_NOT_MULTITHREADED
